# openapi.json
{"openapi":"3.0.2","info":{"title":"Segundo parcial. Ingeniería web","description":"Altair Bueno Calvente. Software 4ºA","version":"0.1.0"},"paths":{"/mensajes":{"get":{"tags":["Mensajes"],"summary":"Get All Mensajes","operationId":"get_all_mensajes_mensajes_get","parameters":[{"required":false,"schema":{"title":"Texto","type":"string"},"name":"texto","in":"query"},{"required":false,"schema":{"title":"Usuario","type":"string"},"name":"usuario","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Mensajes Mensajes Get","type":"array","items":{"$ref":"#/components/schemas/Mensaje"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Mensajes"],"summary":"Create Mensaje","operationId":"create_mensaje_mensajes_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMensaje"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mensajes/{id}":{"get":{"tags":["Mensajes"],"summary":"Get Mensaje","operationId":"get_mensaje_mensajes__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Mensaje"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Mensajes"],"summary":"Update Mensaje","operationId":"update_mensaje_mensajes__id__put","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMensaje"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Mensajes"],"summary":"Delete Mensaje","operationId":"delete_mensaje_mensajes__id__delete","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/usuarios":{"get":{"tags":["Usuarios"],"summary":"Get All Usuarios","operationId":"get_all_usuarios_usuarios_get","parameters":[{"required":false,"schema":{"title":"Alias","type":"string"},"name":"alias","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Usuarios Usuarios Get","type":"array","items":{"$ref":"#/components/schemas/Usuario"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Usuarios"],"summary":"Create Usuario","operationId":"create_usuario_usuarios_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUsuario"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUsuarioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/usuarios/{telefono}":{"get":{"tags":["Usuarios"],"summary":"Get Usuario","operationId":"get_usuario_usuarios__telefono__get","parameters":[{"required":true,"schema":{"title":"Telefono","type":"string"},"name":"telefono","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Usuarios"],"summary":"Update Usuario","operationId":"update_usuario_usuarios__telefono__put","parameters":[{"required":true,"schema":{"title":"Telefono","type":"string"},"name":"telefono","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUsuario"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Usuarios"],"summary":"Delete Usuario","operationId":"delete_usuario_usuarios__telefono__delete","parameters":[{"required":true,"schema":{"title":"Telefono","type":"string"},"name":"telefono","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/usuarios/{telefono}/contactos":{"get":{"tags":["Usuarios"],"summary":"Get Contactos","operationId":"get_contactos_usuarios__telefono__contactos_get","parameters":[{"required":true,"schema":{"title":"Telefono","type":"string"},"name":"telefono","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Contactos Usuarios  Telefono  Contactos Get","type":"array","items":{"$ref":"#/components/schemas/Contacto"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Usuarios"],"summary":"Create Contacto","operationId":"create_contacto_usuarios__telefono__contactos_post","parameters":[{"required":true,"schema":{"title":"Telefono","type":"string"},"name":"telefono","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contacto"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Contacto":{"title":"Contacto","required":["alias","telefono"],"type":"object","properties":{"alias":{"title":"Alias","type":"string"},"telefono":{"title":"Telefono","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Mensaje":{"title":"Mensaje","required":["_id","timestamp","origen","destino","texto"],"type":"object","properties":{"_id":{"title":" Id","type":"string"},"timestamp":{"title":"Timestamp","type":"string","format":"date-time"},"origen":{"title":"Origen","type":"string"},"destino":{"title":"Destino","type":"string"},"texto":{"title":"Texto","maxLength":400,"type":"string"}}},"NewMensaje":{"title":"NewMensaje","required":["origen","destino","texto"],"type":"object","properties":{"origen":{"title":"Origen","type":"string"},"destino":{"title":"Destino","type":"string"},"texto":{"title":"Texto","maxLength":400,"type":"string"}}},"NewUsuario":{"title":"NewUsuario","required":["telefono","alias"],"type":"object","properties":{"telefono":{"title":"Telefono","type":"string"},"alias":{"title":"Alias","type":"string"}}},"NewUsuarioResponse":{"title":"NewUsuarioResponse","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"string"}}},"UpdateMensaje":{"title":"UpdateMensaje","required":["texto"],"type":"object","properties":{"texto":{"title":"Texto","maxLength":400,"type":"string"}}},"UpdateUsuario":{"title":"UpdateUsuario","type":"object","properties":{"alias":{"title":"Alias","type":"string"}}},"Usuario":{"title":"Usuario","required":["telefono","alias"],"type":"object","properties":{"telefono":{"title":"Telefono","type":"string"},"alias":{"title":"Alias","type":"string"},"contactos":{"title":"Contactos","type":"array","items":{"$ref":"#/components/schemas/Contacto"},"default":[]}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"servers":[{"url":"/","description":"Default"},{"url":"http://localhost:8000","description":"Docker compose"}]}